Changed all rpm values to floats so that GenericProviderBase.CalcCMExtraData3 doesn't get invalid cast when calling (filteredData.max_rpm_div10 = rawData.max_rpm_div10 = (float)filteredData.max_rpm / 10.0f;)

+++ GenericTelemetryProvider/BeamNGTelemetryProvider.cs
@@ -124,8 +124,8 @@ namespace GenericTelemetryProvider
         public override void SimulateEngine()
         {
             rawData.max_rpm = telemetryData.max_rpm;
-            rawData.max_gears = telemetryData.max_gears;
-            rawData.gear = telemetryData.gear;
+            rawData.max_gears = (float)telemetryData.max_gears;
+            rawData.gear = (float)telemetryData.gear;
             rawData.idle_rpm = telemetryData.idle_rpm;
             rawData.engine_rate = telemetryData.engine_rate;
